merge randomizing-fish with upgrade-system

+++ Assets/Scripts/MoneyCounter.cs
@@ -7,7 +7,7 @@ public class MoneyCounter : MonoBehaviour
 {
     [SerializeField] private TMP_Text text;
 
-    public int money = 0;
+    public double money = 0f;
 
     void Start()
     {
